Updated regex site name, updated error handling
Prior to this commit the site name could not contain any
periods - like a hostname.  Also, there was an issue where the
error handling was simply evaluating if there was no output from
a command.  During the creation/deletion process there is a
summary that is outputted after the cmdlet exists that should not
cause a failure.  I've updated to explicitly look for an error in
the string output.